ReadmeCopyright.md
All of this code is under copyright in the UK, US and any country in the EU or Europe. Stealing all my code is considered copyright infringement (Unless your not putting it on the internet or not giving all the code to someone,) but using part of my code is allowed and not under the copyright infringement chart. One of the plugins in this code used is made by me and not at all can be downloaded anywhere not even through pip cmd line. As stated on GOV.UK you can not use most of my work. If you're trying to show code to a social platform please give credit to Wiktor. Any other question email me at wiktorswakon@hotmail.com.
All this code is licensed to name: Wiktor Swakon :Under the email of wiktorswakon@hotmail.com...
Official download link on github user Wiky250.